some changes to mgmt command and collection serializer

+++ kpi/management/commands/import_from_url.py
@@ -8,11 +8,11 @@ from optparse import make_option
 
 class Command(BaseCommand):
     option_list = BaseCommand.option_list + (
-        # make_option('--delete',
-        #     action='store_true',
-        #     dest='delete',
-        #     default=False,
-        #     help='Delete all collections for user'),
+        make_option('--destroy',
+            action='store_true',
+            dest='destroy',
+            default=False,
+            help='Delete all collections, assets, and tasks for user'),
         make_option('--username',
             action='store',
             dest='username',
@@ -24,10 +24,10 @@ class Command(BaseCommand):
         if not options.get('username'):
             raise Exception("username flag required '--username'")
         user = User.objects.get(username=options.get('username'))
-        # if options.get('delete'):
-        user.owned_collections.all().delete()
-        user.survey_assets.all().delete()
-        ImportTask.objects.all().delete()
+        if options.get('destroy'):
+            user.owned_collections.all().delete()
+            user.survey_assets.all().delete()
+            ImportTask.objects.filter(user=user).delete()
         url = args[0]
         import_task = ImportTask.objects.create(user=user,
             data={
